refactor(screens): Move settings screen widgets to separate subdirectories
